wip: dedupe seed word ui
